Add README and update TemplateConfig, build.gradle.kts
The README.md file was created with instructions on how to use the Alcosi template service communication library. The TemplateConfig.kts file was updated with modifications to some bean definitions and a new bean for TemplateService was added. Modifications were also made to the build.gradle.kts file to enable a Jar task.