Add package authentication section to README
A detailed section on package authentication has been added to the README file. This outlines how to authenticate packages that require it, even those which are open-source but still require GitHub login. It explains how to set up the build file for this and addresses differences between local builds and CI/CD.